requirements.txt
langchain
langchain_core
langchain_groq
gradio
langchain-google-genai
langchain-openai
python-dotenv
